+++ Mlib/Render/Renderables/Triangle_Sampler/Renderable_Triangle_Sampler.cpp
@@ -220,7 +220,7 @@ void RenderableTriangleSampler::append_sorted_instances_to_queue(
 void RenderableTriangleSampler::extend_aabb(
     const TransformationMatrix<float, double, 3>& mv,
     ExternalRenderPassType render_pass,
-    AxisAlignedBoundingBox<float, 3>& aabb) const
+    AxisAlignedBoundingBox<double, 3>& aabb) const
 {
     // Not yet implemented
 }
