+++ Mlib/Render/Renderables/Triangle_Sampler/Renderable_Triangle_Sampler.hpp
@@ -47,7 +47,7 @@ public:
     virtual void extend_aabb(
         const TransformationMatrix<float, double, 3>& mv,
         ExternalRenderPassType render_pass,
-        AxisAlignedBoundingBox<float, 3>& aabb) const override;
+        AxisAlignedBoundingBox<double, 3>& aabb) const override;
 private:
     const SceneNodeResources& scene_node_resources_;
     const TerrainStyles& terrain_styles_;
